backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
requests>=2.31.0
aiofiles>=23.2.1
PyYAML
